moe: harden the CUTLASS grouped GEMM staging; add MoE GEMM tests
Follow-up to #210, from its review and a new test.

- Staging ring. Each slot now has its own event, recorded behind the GEMM
  that reads it. A slot is reused only once that event completes, and the
  wait polls, as stream_wait_spin does. Before, a generation event recorded
  before the last GEMM made the host drain its GPU every 256 calls, and a
  stream change synchronized the previous stream. That handle may already
  be destroyed: the new test hit "context is destroyed" when a stream was
  recreated on the same thread, and v1.5.3's pointer ring fails the same
  way. Rings are now per thread and per device. All slots come from one
  pinned and one device allocation (32 KB per slot, grown geometrically),
  instead of 512 allocations and step-wise regrowth above 272 experts.
- Fallback. What cublasGemmEx rejects (dimension <= 0, leading dimension
  below the extent) goes to the cublasGemmEx loop, so the caller gets
  cuBLAS's error instead of a bad launch. CUBLAS_OP_C counts as
  transposed. A bf16 call that falls back is reported once on stderr.
- Pointer offsets in the expert GEMM wrappers are computed in size_t;
  weight_idx * M * K overflowed int for DeepSeek-V3-sized experts.
- Stale comments about cuBLAS grouped GEMM are updated.
- testing/kernels/test-moe-grouped-gemm.cu:
  - Every bf16 entry point matches a double-precision reference.
  - A token's output does not depend on its expert's token count.
  - With two GPUs, while one worker is blocked in a launch inside
    moe_grouped_gemm, the other worker's call still returns. v1.5.3's
    library fails this test (cuBLAS grouped GEMM's lock); this build
    passes it.

+++ csrc/src/kernels/moe/moe_grouped_gemm_backward.cu
@@ -186,9 +186,9 @@ void moe_grouped_gemm_down_backward_impl(
 
         const int weight_idx = weight_is_compact ? e : global_idx;
         A_vec.push_back(weight_ptrs ? static_cast<const T*>(weight_ptrs[weight_idx])
-                                    : weights + weight_idx * hidden_size * intermediate_size);
-        B_vec.push_back(d_output + h_offsets[global_idx] * hidden_size);
-        C_vec.push_back(d_input + h_offsets[global_idx] * intermediate_size);
+                                    : weights + static_cast<std::size_t>(weight_idx) * hidden_size * intermediate_size);
+        B_vec.push_back(d_output + static_cast<std::size_t>(h_offsets[global_idx]) * hidden_size);
+        C_vec.push_back(d_input + static_cast<std::size_t>(h_offsets[global_idx]) * intermediate_size);
     }
 
     if (m_vec.empty()) return;
@@ -281,9 +281,9 @@ void moe_grouped_gemm_gate_up_backward_impl(
 
         const int weight_idx = weight_is_compact ? e : global_idx;
         A_vec.push_back(weight_ptrs ? static_cast<const T*>(weight_ptrs[weight_idx])
-                                    : weights + weight_idx * gate_up_dim * hidden_size);
-        B_vec.push_back(d_gate_up + h_offsets[global_idx] * gate_up_dim);
-        C_vec.push_back(d_input + h_offsets[global_idx] * hidden_size);
+                                    : weights + static_cast<std::size_t>(weight_idx) * gate_up_dim * hidden_size);
+        B_vec.push_back(d_gate_up + static_cast<std::size_t>(h_offsets[global_idx]) * gate_up_dim);
+        C_vec.push_back(d_input + static_cast<std::size_t>(h_offsets[global_idx]) * hidden_size);
     }
 
     if (m_vec.empty()) return;
@@ -375,9 +375,9 @@ void moe_grouped_gemm_up_backward_impl(
 
         const int weight_idx = weight_is_compact ? e : global_idx;
         A_vec.push_back(weight_ptrs ? static_cast<const T*>(weight_ptrs[weight_idx])
-                                    : weights + weight_idx * up_dim * hidden_size);
-        B_vec.push_back(d_up + h_offsets[global_idx] * up_dim);
-        C_vec.push_back(d_input + h_offsets[global_idx] * hidden_size);
+                                    : weights + static_cast<std::size_t>(weight_idx) * up_dim * hidden_size);
+        B_vec.push_back(d_up + static_cast<std::size_t>(h_offsets[global_idx]) * up_dim);
+        C_vec.push_back(d_input + static_cast<std::size_t>(h_offsets[global_idx]) * hidden_size);
     }
 
     if (m_vec.empty()) return;
